requirements.txt
Pillow==9.5.0
google-cloud-aiplatform==1.27.1
clarifai-grpc==9.6.0